correción controller
