+++ middleware/is_active.js
@@ -1,4 +1,4 @@
-import Author from '../../models/Author.js';
+import Author from '../models/Author.js';
 
 const is_active = async (req, res, next) => {
   try {
